longlong

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: sivinska <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 17:30:29 by sivinska          #+#    #+#             */
-/*   Updated: 2018/04/25 13:25:57 by sivinska         ###   ########.fr       */
+/*   Updated: 2018/04/25 15:26:05 by sivinska         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,8 +85,8 @@ void				ft_lstdel(t_list **alst, void (*del)(void *, size_t));
 void				ft_lstadd(t_list **alst, t_list *new);
 void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
-char				*ft_convert_to_base(unsigned int nbr, int base);
-char				*ft_convert_to_base_long(long int nbr, int base);
+char				*ft_convert_to_base(unsigned long long nbr, int base);
+char				*ft_convert_to_base_long(unsigned long int nbr, int base);
 char				*ft_str_to_upper(char *str);
 char				*ft_strrev(char *str);
 int					ft_get_nbr_len(int n);
